# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Campus Companion API
  version: 0.4.0-m1-baseline-freeze
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: >-
    校园同行信息撮合 API。契约只覆盖学生认证、固定路线、乘客组队、
    信息服务费和联系方式授权，不包含司机、车辆、运价或运输履约。
servers:
  - url: https://api.example.invalid/v1
    description: Placeholder; M6 才配置真实域名
...
security:
  - bearerAuth: []
...
paths:
...
  /auth/wechat/login:
    post:
      tags: [Auth]
      security: []
      operationId: wechatLogin
      summary: 使用一次性 wx.login code 换取平台会话
      description: 同一个微信 code 只能消费一次；重放不得创建第二个会话。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WechatLoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '400': { $ref: '#/components/responses/ValidationError' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /auth/refresh:
    post:
      tags: [Auth]
      security: []
      operationId: refreshSession
      summary: 轮换刷新令牌
      description: Refresh Token family 单次轮换；并发刷新只允许一个请求成功，检测到已轮换令牌重放时撤销该 family。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [refreshToken]
              properties:
                refreshToken: { type: string, minLength: 32, maxLength: 2048 }
      responses:
        '200':
          description: 刷新成功，旧令牌立即失效
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/logout:
    post:
      tags: [Auth]
      summary: Revoke the current platform session
      operationId: logout
      description: 对同一可识别会话重复调用安全返回，不恢复或新建会话。
      responses:
        '204': { description: 当前会话已撤销 }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /me:
    get:
      tags: [Auth]
      summary: Get the authenticated user profile
      operationId: getMe
      responses:
        '200':
          description: 当前用户（不返回联系方式明文）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Me' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    delete:
      tags: [Auth]
      summary: Request account deletion
      operationId: requestAccountDeletion
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '202': { description: 注销任务已受理 }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /verifications:
    post:
      tags: [Verification]
      summary: Create a student verification draft
      operationId: createVerification
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateVerificationRequest' }
      responses:
        '201':
          description: 已创建申请和短期私有上传凭证
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VerificationUploadResponse' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
    get:
      tags: [Verification]
      summary: Get the current student verification
      operationId: getCurrentVerification
      responses:
        '200':
          description: 当前申请状态；不返回材料 URL
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Verification' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /verifications/{verificationId}/submit:
    post:
      tags: [Verification]
      summary: Submit a verification draft for manual review
      operationId: submitVerification
      description: >-
        服务端对自己生成的对象键执行HEAD检查，验证大小、类型和摘要后进入审核队列。
        首次提交进入PENDING并同时写入submittedAt/latestSubmittedAt；补交进入
        RESUBMISSION_PENDING，仅更新latestSubmittedAt并保留原submittedAt/reviewedAt。
      parameters:
        - name: verificationId
          in: path
          required: true
          schema: { $ref: '#/components/schemas/Uuid' }
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [uploads]
              properties:
                uploads:
                  type: array
                  minItems: 1
                  maxItems: 2
                  uniqueItems: true
                  items:
                    type: object
                    additionalProperties: false
                    required: [type, uploadEtag]
                    properties:
                      type: { $ref: '#/components/schemas/VerificationAssetType' }
                      uploadEtag: { type: string, pattern: '^[A-Fa-f0-9]{64}$' }
      responses:
        '202':
          description: 材料已验证并进入人工审核队列
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Verification' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /verifications/{verificationId}/resubmission-upload:
    post:
      tags: [Verification]
      summary: Create a replacement upload credential after a resubmission decision
      operationId: createResubmissionUpload
      description: >-
        仅允许REQUIRE_RESUBMISSION进入RESUBMISSION_AWAITING_UPLOAD；旧材料立即失效并进入删除，
        原submittedAt、latestSubmittedAt和reviewedAt不得被清空或改写。
      parameters:
        - name: verificationId
          in: path
          required: true
          schema: { $ref: '#/components/schemas/Uuid' }
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [evidenceTypes]
              properties:
                evidenceTypes:
                  type: array
                  minItems: 1
                  maxItems: 2
                  uniqueItems: true
                  items: { $ref: '#/components/schemas/VerificationAssetType' }
      responses:
        '201':
          description: 已创建补交上传凭证并保留历史审核时间
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VerificationUploadResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /admin/auth/logout:
    post:
      tags: [Admin]
      security:
        - adminCookie: []
          csrfToken: []
      summary: Revoke the current administrator session
      operationId: adminLogout
      description: 重复调用安全收敛到会话已撤销；不得恢复旧 Cookie 或 CSRF Token。
      responses:
        '204': { description: 管理员会话已撤销 }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /admin/verifications/{verificationId}:
    get:
      tags: [Admin]
      security:
        - adminCookie: []
          csrfToken: []
      summary: Get campus-scoped verification metadata without asset addresses
      operationId: getAdminVerification
      x-required-roles: [VERIFICATION_REVIEWER]
      x-campus-scope: required
      parameters:
        - name: verificationId
          in: path
          required: true
          schema: { $ref: '#/components/schemas/Uuid' }
      responses:
        '200':
          description: 脱敏审核元数据
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminVerification' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminCookie:
      type: apiKey
      in: cookie
      name: __Host-admin_session
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: >-
        绑定认证主体、operationId 与规范化请求摘要并保留24小时；同键同摘要返回首次结果，
        同键不同摘要返回409 IDEMPOTENCY_CONFLICT。不得用于登录、刷新、登出或微信回调。
      schema: { type: string, minLength: 16, maxLength: 128 }
...
  responses:
    ValidationError:
      description: 输入不合法
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: 未登录或会话失效
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: 认证、权限、资源归属或风控检查失败
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: 资源不存在或为防枚举而隐藏
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: 当前状态不允许操作或幂等键冲突
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    RateLimited:
      description: 请求频率受限
      headers:
        Retry-After:
          schema: { type: integer, minimum: 1 }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    Uuid:
      type: string
      format: uuid
    WechatLoginRequest:
      type: object
      additionalProperties: false
      required: [code]
      properties:
        code: { type: string, minLength: 1, maxLength: 128 }
    SessionResponse:
      type: object
      additionalProperties: false
      required: [accessToken, refreshToken, expiresInSeconds, user]
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        expiresInSeconds: { type: integer, minimum: 60, maximum: 3600 }
        user: { $ref: '#/components/schemas/Me' }
    Me:
      type: object
      additionalProperties: false
      required: [id, campusId, accountStatus, verificationStatus, genderDeclaration]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        campusId: { $ref: '#/components/schemas/Uuid' }
        accountStatus: { type: string, enum: [ACTIVE, RESTRICTED, DELETION_PENDING] }
        verificationStatus:
          oneOf:
            - { type: string, const: NOT_SUBMITTED }
            - { $ref: '#/components/schemas/VerificationStatus' }
        genderDeclaration: { type: string, enum: [MALE, FEMALE, UNDISCLOSED] }
        hasWechatContact: { type: boolean }
    CreateVerificationRequest:
      type: object
      additionalProperties: false
      required: [campusId, studentNumber, genderDeclaration, sensitiveInfoConsentVersion, evidenceTypes]
      properties:
        campusId: { $ref: '#/components/schemas/Uuid' }
        studentNumber: { type: string, minLength: 4, maxLength: 64, writeOnly: true }
        genderDeclaration: { type: string, enum: [MALE, FEMALE, UNDISCLOSED] }
        sensitiveInfoConsentVersion: { type: string, minLength: 1, maxLength: 50 }
        evidenceTypes:
          type: array
          minItems: 1
          maxItems: 2
          uniqueItems: true
          items: { $ref: '#/components/schemas/VerificationAssetType' }
    VerificationAssetType:
      type: string
      enum: [STUDENT_CARD, WECOM_SCREENSHOT]
    VerificationUploadResponse:
      type: object
      additionalProperties: false
      required: [verification, uploads]
      properties:
        verification: { $ref: '#/components/schemas/VerificationUploadReady' }
        uploads:
          type: array
          minItems: 1
          maxItems: 2
          items:
            type: object
            additionalProperties: false
            required: [type, uploadUrl, uploadExpiresAt]
            properties:
              type: { $ref: '#/components/schemas/VerificationAssetType' }
              uploadUrl: { type: string, format: uri, readOnly: true }
              uploadExpiresAt: { type: string, format: date-time }
    VerificationStatus:
      type: string
      enum:
        - AWAITING_UPLOAD
        - UPLOAD_EXPIRED
        - PENDING
        - VERIFIED
        - REJECTED
        - REQUIRE_RESUBMISSION
        - RESUBMISSION_AWAITING_UPLOAD
        - RESUBMISSION_PENDING
        - VERIFICATION_EXPIRED
    VerificationUploadReady:
      oneOf:
        - $ref: '#/components/schemas/VerificationAwaitingUpload'
        - $ref: '#/components/schemas/VerificationResubmissionAwaitingUpload'
      discriminator:
        propertyName: status
        mapping:
          AWAITING_UPLOAD: '#/components/schemas/VerificationAwaitingUpload'
          RESUBMISSION_AWAITING_UPLOAD: '#/components/schemas/VerificationResubmissionAwaitingUpload'
    Verification:
      oneOf:
        - $ref: '#/components/schemas/VerificationAwaitingUpload'
        - $ref: '#/components/schemas/VerificationUploadExpired'
        - $ref: '#/components/schemas/VerificationPending'
        - $ref: '#/components/schemas/VerificationVerified'
        - $ref: '#/components/schemas/VerificationRejected'
        - $ref: '#/components/schemas/VerificationRequiresResubmission'
        - $ref: '#/components/schemas/VerificationCredentialExpired'
        - $ref: '#/components/schemas/VerificationResubmissionAwaitingUpload'
        - $ref: '#/components/schemas/VerificationResubmissionPending'
      discriminator:
        propertyName: status
        mapping:
          AWAITING_UPLOAD: '#/components/schemas/VerificationAwaitingUpload'
          UPLOAD_EXPIRED: '#/components/schemas/VerificationUploadExpired'
          PENDING: '#/components/schemas/VerificationPending'
          VERIFIED: '#/components/schemas/VerificationVerified'
          REJECTED: '#/components/schemas/VerificationRejected'
          REQUIRE_RESUBMISSION: '#/components/schemas/VerificationRequiresResubmission'
          RESUBMISSION_AWAITING_UPLOAD: '#/components/schemas/VerificationResubmissionAwaitingUpload'
          RESUBMISSION_PENDING: '#/components/schemas/VerificationResubmissionPending'
          VERIFICATION_EXPIRED: '#/components/schemas/VerificationCredentialExpired'
    VerificationAwaitingUpload:
      type: object
      additionalProperties: false
      required: [id, campusId, studentNumberLast4, status, evidenceTypes, submittedAt, latestSubmittedAt, reviewedAt, expiresAt, reasonCode]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        campusId: { $ref: '#/components/schemas/Uuid' }
        studentNumberLast4: { type: string, pattern: '^[A-Za-z0-9]{4}$' }
        status: { type: string, const: AWAITING_UPLOAD }
        evidenceTypes: { type: array, minItems: 1, maxItems: 2, uniqueItems: true, items: { $ref: '#/components/schemas/VerificationAssetType' } }
        submittedAt: { type: 'null' }
        latestSubmittedAt: { type: 'null' }
        reviewedAt: { type: 'null' }
        expiresAt: { type: 'null' }
        reasonCode: { type: 'null' }
      example:
        id: 00000000-0000-0000-0000-000000000101
        campusId: 00000000-0000-0000-0000-000000000001
        studentNumberLast4: '1234'
        status: AWAITING_UPLOAD
        evidenceTypes: [STUDENT_CARD]
        submittedAt: null
        latestSubmittedAt: null
        reviewedAt: null
        expiresAt: null
        reasonCode: null
    VerificationUploadExpired:
      type: object
      additionalProperties: false
      required: [id, campusId, studentNumberLast4, status, evidenceTypes, submittedAt, latestSubmittedAt, reviewedAt, expiresAt, reasonCode]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        campusId: { $ref: '#/components/schemas/Uuid' }
        studentNumberLast4: { type: string, pattern: '^[A-Za-z0-9]{4}$' }
        status: { type: string, const: UPLOAD_EXPIRED }
        evidenceTypes: { type: array, minItems: 1, maxItems: 2, uniqueItems: true, items: { $ref: '#/components/schemas/VerificationAssetType' } }
        submittedAt: { type: 'null' }
        latestSubmittedAt: { type: 'null' }
        reviewedAt: { type: 'null' }
        expiresAt: { type: 'null' }
        reasonCode: { type: string, const: UPLOAD_CREDENTIAL_EXPIRED }
    VerificationPending:
      type: object
      additionalProperties: false
      required: [id, campusId, studentNumberLast4, status, evidenceTypes, submittedAt, latestSubmittedAt, reviewedAt, expiresAt, reasonCode]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        campusId: { $ref: '#/components/schemas/Uuid' }
        studentNumberLast4: { type: string, pattern: '^[A-Za-z0-9]{4}$' }
        status: { type: string, const: PENDING }
        evidenceTypes: { type: array, minItems: 1, maxItems: 2, uniqueItems: true, items: { $ref: '#/components/schemas/VerificationAssetType' } }
        submittedAt: { type: string, format: date-time }
        latestSubmittedAt: { type: string, format: date-time }
        reviewedAt: { type: 'null' }
        expiresAt: { type: 'null' }
        reasonCode: { type: 'null' }
    VerificationVerified:
      type: object
      additionalProperties: false
      required: [id, campusId, studentNumberLast4, status, evidenceTypes, submittedAt, latestSubmittedAt, reviewedAt, expiresAt, reasonCode]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        campusId: { $ref: '#/components/schemas/Uuid' }
        studentNumberLast4: { type: string, pattern: '^[A-Za-z0-9]{4}$' }
        status: { type: string, const: VERIFIED }
        evidenceTypes: { type: array, minItems: 1, maxItems: 2, uniqueItems: true, items: { $ref: '#/components/schemas/VerificationAssetType' } }
        submittedAt: { type: string, format: date-time }
        latestSubmittedAt: { type: string, format: date-time }
        reviewedAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
        reasonCode: { type: 'null' }
    VerificationRejected:
      type: object
      additionalProperties: false
      required: [id, campusId, studentNumberLast4, status, evidenceTypes, submittedAt, latestSubmittedAt, reviewedAt, expiresAt, reasonCode]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        campusId: { $ref: '#/components/schemas/Uuid' }
        studentNumberLast4: { type: string, pattern: '^[A-Za-z0-9]{4}$' }
        status: { type: string, const: REJECTED }
        evidenceTypes: { type: array, minItems: 1, maxItems: 2, uniqueItems: true, items: { $ref: '#/components/schemas/VerificationAssetType' } }
        submittedAt: { type: string, format: date-time }
        latestSubmittedAt: { type: string, format: date-time }
        reviewedAt: { type: string, format: date-time }
        expiresAt: { type: 'null' }
        reasonCode: { type: string, minLength: 1, maxLength: 100 }
    VerificationRequiresResubmission:
      type: object
      additionalProperties: false
      required: [id, campusId, studentNumberLast4, status, evidenceTypes, submittedAt, latestSubmittedAt, reviewedAt, expiresAt, reasonCode]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        campusId: { $ref: '#/components/schemas/Uuid' }
        studentNumberLast4: { type: string, pattern: '^[A-Za-z0-9]{4}$' }
        status: { type: string, const: REQUIRE_RESUBMISSION }
        evidenceTypes: { type: array, minItems: 1, maxItems: 2, uniqueItems: true, items: { $ref: '#/components/schemas/VerificationAssetType' } }
        submittedAt: { type: string, format: date-time }
        latestSubmittedAt: { type: string, format: date-time }
        reviewedAt: { type: string, format: date-time }
        expiresAt: { type: 'null' }
        reasonCode: { type: string, minLength: 1, maxLength: 100 }
    VerificationCredentialExpired:
      type: object
      additionalProperties: false
      required: [id, campusId, studentNumberLast4, status, evidenceTypes, submittedAt, latestSubmittedAt, reviewedAt, expiresAt, reasonCode]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        campusId: { $ref: '#/components/schemas/Uuid' }
        studentNumberLast4: { type: string, pattern: '^[A-Za-z0-9]{4}$' }
        status: { type: string, const: VERIFICATION_EXPIRED }
        evidenceTypes: { type: array, minItems: 1, maxItems: 2, uniqueItems: true, items: { $ref: '#/components/schemas/VerificationAssetType' } }
        submittedAt: { type: string, format: date-time }
        latestSubmittedAt: { type: string, format: date-time }
        reviewedAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
        reasonCode: { type: string, const: CREDENTIAL_EXPIRED }
    VerificationResubmissionAwaitingUpload:
      type: object
      additionalProperties: false
      required: [id, campusId, studentNumberLast4, status, evidenceTypes, submittedAt, latestSubmittedAt, reviewedAt, expiresAt, reasonCode]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        campusId: { $ref: '#/components/schemas/Uuid' }
        studentNumberLast4: { type: string, pattern: '^[A-Za-z0-9]{4}$' }
        status: { type: string, const: RESUBMISSION_AWAITING_UPLOAD }
        evidenceTypes: { type: array, minItems: 1, maxItems: 2, uniqueItems: true, items: { $ref: '#/components/schemas/VerificationAssetType' } }
        submittedAt: { type: string, format: date-time }
        latestSubmittedAt: { type: string, format: date-time }
        reviewedAt: { type: string, format: date-time }
        expiresAt: { type: 'null' }
        reasonCode: { type: string, minLength: 1, maxLength: 100 }
    VerificationResubmissionPending:
      type: object
      additionalProperties: false
      required: [id, campusId, studentNumberLast4, status, evidenceTypes, submittedAt, latestSubmittedAt, reviewedAt, expiresAt, reasonCode]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        campusId: { $ref: '#/components/schemas/Uuid' }
        studentNumberLast4: { type: string, pattern: '^[A-Za-z0-9]{4}$' }
        status: { type: string, const: RESUBMISSION_PENDING }
        evidenceTypes: { type: array, minItems: 1, maxItems: 2, uniqueItems: true, items: { $ref: '#/components/schemas/VerificationAssetType' } }
        submittedAt: { type: string, format: date-time }
        latestSubmittedAt: { type: string, format: date-time }
        reviewedAt: { type: string, format: date-time }
        expiresAt: { type: 'null' }
        reasonCode: { type: [string, 'null'], maxLength: 100 }
...
    AdminVerification:
      type: object
      additionalProperties: false
      required: [id, campusId, studentNumberLast4, status, submittedAt, latestSubmittedAt, reviewedAt, availableAssetTypes]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        campusId: { $ref: '#/components/schemas/Uuid' }
        studentNumberLast4: { type: string, pattern: '^[A-Za-z0-9]{4}$' }
        status: { $ref: '#/components/schemas/VerificationStatus' }
        submittedAt: { type: [string, 'null'], format: date-time }
        latestSubmittedAt: { type: [string, 'null'], format: date-time }
        reviewedAt: { type: [string, 'null'], format: date-time }
        availableAssetTypes:
          type: array
          maxItems: 2
          uniqueItems: true
          items: { $ref: '#/components/schemas/VerificationAssetType' }
        reasonCode: { type: [string, 'null'], maxLength: 100 }
...
    FieldViolation:
      type: object
      additionalProperties: false
      required: [field, constraint]
      properties:
        field: { type: string, minLength: 1, maxLength: 100 }
        constraint: { type: string, minLength: 1, maxLength: 100 }
    ErrorDetails:
      type: object
      additionalProperties: false
      properties:
        field: { type: string, minLength: 1, maxLength: 100 }
        constraint: { type: string, minLength: 1, maxLength: 100 }
        violations:
          type: array
          maxItems: 20
          items: { $ref: '#/components/schemas/FieldViolation' }
        retryAfterSeconds: { type: integer, minimum: 1, maximum: 86400 }
        currentVersion: { type: integer, minimum: 0 }
        expectedVersion: { type: integer, minimum: 0 }
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, requestId]
          properties:
            code:
              type: string
              enum:
                - VALIDATION_ERROR
                - AUTH_REQUIRED
                - SESSION_EXPIRED
                - STUDENT_NOT_VERIFIED
                - ACCOUNT_RESTRICTED
                - RESOURCE_NOT_FOUND
                - RESOURCE_FORBIDDEN
                - IDEMPOTENCY_CONFLICT
                - OVERLAPPING_ACTIVE_GROUP
                - GROUP_NOT_JOINABLE
                - GROUP_CAPACITY_EXCEEDED
                - GENDER_PREFERENCE_INCOMPATIBLE
                - FORMATION_NOT_READY
                - FORMATION_EXPIRED
                - PAYMENT_NOT_CONFIRMED
                - CONTACTS_NOT_UNLOCKED
                - CONTACT_CONSENT_REQUIRED
                - CONTACT_CONSENT_VERSION_MISMATCH
                - CONTACT_CONSENT_REVOKED
                - GROUP_REFUND_IN_PROGRESS
                - REFUND_NOT_ELIGIBLE
                - ADMIN_CSRF_INVALID
                - ADMIN_ROLE_REQUIRED
                - ADMIN_CAMPUS_FORBIDDEN
                - ADMIN_REAUTH_REQUIRED
                - VERIFICATION_ASSET_INVALID
                - VERIFICATION_UPLOAD_EXPIRED
                - VERIFICATION_RESUBMISSION_REQUIRED
                - VERIFICATION_ASSET_GRANT_UNAVAILABLE
                - RATE_LIMITED
                - INTERNAL_ERROR
            message: { type: string, maxLength: 200 }
            requestId: { type: string, maxLength: 100 }
            details: { $ref: '#/components/schemas/ErrorDetails' }